Handle cyclic import in _have_debugger()
This function may cause an import cycle when it is called during early
startup.
Specifically, during import of debugger.py, the settings are loaded.
Exceptions during load are forwarded to the logging system, where they
are handled by TerminalOrStreamHandler. This is turn will call
_have_debugger() to determine whether to display an UI dialog.
The function attempts to import the debugger.py, completing the import cycle.